add force push option

+++ shell/push.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
-cd ../webUI/git/$1
-git push -u https://$3@github.com/$2/$1 --force
-sleep 1
+if [ $4 -eq "1" ]; then
+    cd ../webUI/git/$1
+    git push -u https://$3@github.com/$2/$1 --force
+    sleep 1
+else
+    cd ../webUI/git/$1
+    git push -u https://$3@github.com/$2/$1 
+    sleep 1
+fi
